feat: add Makefile and run_experiments_env.yml
# Context
Previously, there was no real declarative way to define/install
runtime environments for executing experiments in this
repository.

# Proposed feature
This commit adds a `.yml` file with a dependency list for the
environment to run experiments in. It further introduces
a Make entrypoint, which installs that environment and adds
the source code of following packages to the environment:
- gumibair
- experiments/gumibair_experiments

# Future work
There is currently no runtime environment to fetch
and preprocess the input data, using the source code
in `data_pipeline`. A future commit should either
add the required dependencies for that to `run_experiments_env.yml`
or create a new environment for that.