refactor: fetch and curate data from the root directory
# Context
Previously, the data was fetched and curated from the
`data_pipeline/data` directory, with its own Makefile.
Also, there was no declarative way to install the dependencies
for fetching and curating the data.

# Proposed changes
This commit introduces a new environment to fetch and curate the data.
Further, it applies some changes to the R script that
fetches the data, so that it only fetches the marker  abundance data.
Additionally, the commit adds some rules to the top-level Makefile to
fetch and curate the data from the root directory & removes the Makefile
inside the `data_pipeline` directory.